Return early when result is empty in verify_value
When the result is empty in an assertion, `verify_value` returns 0, which causes the test to succeed when it should fail. The test I used for this was 
``` 
run verify "'status' is 'running' for pods named 'app'"
```

When that pod doesn't exist prior to this change, the test would succeed.